day32/streamsocial-source-connector/requirements.txt
kafka-python==2.0.2
confluent-kafka==2.4.0
watchdog==4.0.1
flask==3.0.3
flask-cors==4.0.1
prometheus-client==0.20.0
PyYAML==6.0.1
psutil==5.9.8
requests==2.31.0
boto3==1.34.131
python-dotenv==1.0.1
jsonschema==4.22.0
retry==0.9.2
click==8.1.7
gunicorn==22.0.0
pytest==8.2.2
pytest-asyncio==0.23.7
coverage==7.5.3
black==24.4.2
flake8==7.1.0
